+++ Scripts/UnityPython.cs
@@ -2,13 +2,16 @@
 using Microsoft.Scripting.Hosting;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Reflection;
+using UnityEngine;
 
 /// <summary>
 /// Convenience class for creating a Python engine integrated with Unity.
 ///
 /// All scripts executed by the ScriptEngine created by this class will:
+/// * Redirect output to Unity's console
 /// * Be able to import any class in a `UnityEngine*` namespace
 /// * Be able to import any class in a `UnityEditor*` namespace, if the script
 ///   is running in the UnityEditor
@@ -18,6 +21,16 @@ public static class UnityPython
 	public static ScriptEngine CreateEngine(IDictionary<string, object> options = null)
 	{
 		var engine = Python.CreateEngine(options);
+
+		// Redirect IronPython IO
+		var infoStream = new MemoryStream();
+		var infoWriter = new UnityLogWriter(Debug.Log, infoStream);
+		engine.Runtime.IO.SetOutput(infoStream, infoWriter);
+
+		var errorStream = new MemoryStream();
+		var errorWriter = new UnityLogWriter(Debug.Log, errorStream);
+		engine.Runtime.IO.SetErrorOutput(errorStream, errorWriter);
+
 
 		// Load assemblies for the `UnityEngine*` namespaces
 		foreach (var assembly in GetAssembliesInNamespace("UnityEngine"))